requirements.txt
# Instalar con: pip install -r requirements.txt

discord.py[voice]
python-dotenv
yt-dlp
PyNaCl